# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TempEdge API
  description: |
    Temperature weather prediction market trading system.
    Microservice architecture with 6 HTTP services communicating via REST APIs.

    **Services:**
    - **data-svc** (3005) — Database, sessions, trades, config
    - **trading-svc** (3004) — Buy, sell, redeem via Polymarket CLOB
    - **weather-svc** (3002) — Forecast + current conditions
    - **market-svc** (3003) — Polymarket event discovery
    - **liquidity-svc** (3001) — WebSocket order book streaming
    - **dashboard-svc** (3000) — Admin UI + API aggregation
  version: 1.0.0
  contact:
    name: TempEdge
servers:
  - url: http://localhost:3005
    description: data-svc (local)
  - url: http://localhost:3004
    description: trading-svc (local)
  - url: http://localhost:3002
    description: weather-svc (local)
  - url: http://localhost:3003
    description: market-svc (local)
  - url: http://localhost:3001
    description: liquidity-svc (local)
  - url: http://localhost:3000
    description: dashboard-svc (local)
...
components:
  schemas:
...
    AlertInput:
      type: object
      required: [sessionId, timestamp, type]
      properties:
        sessionId:
          type: string
          minLength: 1
        timestamp:
          type: string
          minLength: 1
        type:
          type: string
          enum: [forecast_shift, range_shift, price_spike, buy_executed, phase_change, sell_executed, redeem_executed]
        message:
          type: string
        data:
          type: object
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
paths:
...
  /api/sessions:
    get:
      tags: [Sessions]
      summary: List all sessions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Sessions ordered by target_date DESC
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
...
  /api/sessions/{id}:
    get:
      tags: [Sessions]
      summary: Get session by market+date
      parameters:
        - name: id
          in: path
          required: true
          description: Target date (YYYY-MM-DD)
          schema:
            type: string
      responses:
        '200':
          description: Session found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Sessions]
      summary: Upsert session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionInput'
      responses:
        '200':
          description: Session upserted
          content:
            application/json:
              schema:
                type: object
                properties:
                  upserted:
                    type: boolean
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Sessions]
      summary: Update session fields
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, completed, stopped]
                phase:
                  type: string
                  enum: [scout, track, buy, monitor, resolve]
      responses:
        '200':
          description: Session updated
        '400':
          description: Validation error
  /api/db/sessions:
    post:
      tags: [Sessions]
      summary: Upsert session (alternate endpoint)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionInput'
      responses:
        '201':
          description: Session upserted
          content:
            application/json:
              schema:
                type: object
                properties:
                  upserted:
                    type: boolean
                  existingId:
                    type: string
                    nullable: true
        '400':
          description: Validation error
        '409':
          description: Conflict (FK constraint)

  # Trades
  /api/trades:
    get:
      tags: [Trades]
      summary: List trades
      parameters:
        - name: date
          in: query
          description: Filter by target date
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Trades list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trade'
    post:
      tags: [Trades]
      summary: Insert a new trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradeInput'
      responses:
        '201':
          description: Trade created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
        '400':
          description: Validation error
  /api/trades/{id}:
    patch:
      tags: [Trades]
      summary: Update trade status/verification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [placed, filled, failed, redeemed]
                verifiedAt:
                  type: string
                actualCost:
                  type: number
                  minimum: 0
                totalProceeds:
                  type: number
                  minimum: 0
      responses:
        '200':
          description: Trade updated
        '400':
          description: Validation error
  /api/trades/log:
    get:
      tags: [Trades]
      summary: Trade log with positions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Trades with joined position data
...
  /api/trade-summary:
    get:
      tags: [Trades]
      summary: Rich trade summary for dashboard (session files + DB)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 15
      responses:
        '200':
          description: Merged trade summary from session files and database
          content:
            application/json:
              schema:
                type: object
                properties:
                  trades:
                    type: array
                    items:
                      type: object

  # Positions
  /api/positions:
    post:
      tags: [Positions]
      summary: Batch insert positions for a trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionsInsertInput'
      responses:
        '201':
          description: Positions inserted
          content:
            application/json:
              schema:
                type: object
                properties:
                  inserted:
                    type: integer
        '400':
          description: Validation error
  /api/positions/active:
    get:
      tags: [Positions]
      summary: Get active (non-sold, non-failed) positions
      parameters:
        - name: date
          in: query
          required: true
          description: Target date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Active positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'
  /api/positions/{id}:
    get:
      tags: [Positions]
      summary: Get position by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Position found
        '404':
          description: Position not found
    patch:
      tags: [Positions]
      summary: Update position fields
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Position updated
  /api/positions/{id}/sold:
    patch:
      tags: [Positions]
      summary: Mark position as sold
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sellPrice, soldAt]
              properties:
                sellPrice:
                  type: number
                  minimum: 0
                soldAt:
                  type: string
                  format: date-time
                sellOrderId:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Position marked as sold
        '400':
          description: Validation error
  /api/positions/{id}/redeemed:
    patch:
      tags: [Positions]
      summary: Mark position as redeemed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [redeemedValue, redeemedAt]
              properties:
                redeemedValue:
                  type: number
                  minimum: 0
                redeemedAt:
                  type: string
                  format: date-time
                redeemedTx:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Position marked as redeemed
        '400':
          description: Validation error

  # Snapshots
  /api/snapshots:
    post:
      tags: [Snapshots]
      summary: Insert monitoring snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnapshotInput'
      responses:
        '201':
          description: Snapshot inserted
        '400':
          description: Validation error
        '409':
          description: Foreign key conflict
  /api/snapshots/{sessionId}:
    get:
      tags: [Snapshots]
      summary: Get snapshots for chart rendering
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 500
      responses:
        '200':
          description: Snapshots ordered by timestamp ASC

  # Alerts
  /api/alerts:
    post:
      tags: [Alerts]
      summary: Insert system alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertInput'
      responses:
        '201':
          description: Alert inserted
        '400':
          description: Validation error
        '409':
          description: Foreign key conflict

  # Spend
...